refactor: move schemas into their own files
